Added a wrapper for ALVAR 1.2.0 and a description for compiling this version in the Installation Guide.
The Scene.cs has been fixed so that Tutorial 8 works fine now with ALVAR. For ARTag, due to some structural changes for eliminating marker update delay, it has to be run with State.IsMultiCore set to false. Otherwise, you won't see anything overlaid due to Exception which is caused by a conflict of accessing openGL matrix in multi-threaded code.

git-svn-id: https://goblinxna.svn.codeplex.com/svn@30921 a83551a4-30f6-4d81-a974-c6ced450ddbf